pyproject.toml
[tool.poetry]
name = "test-deepl"
version = "0.1.0"
description = ""
authors = ["Fabio Putino"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.4"
torch = "^2.1.2"
torchvision = "^0.16.2"
kornia = "^0.7.1"
lightning = "^2.1.3"
PyYAML = "^6.0.1"
wandb = "^0.16.2"
numpy = "^1.26.3"
seaborn = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
